+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Fixed
 * Ozon Content: `wijzigactie` van `opschrift` in `kop` toont geen renvooi ([#3413](https://github.com/dso-toolkit/dso-toolkit/issues/3413))
 
+### Fixed
+* Ozon Content: `wijzigactie` van `opschrift` in `kop` toont geen renvooi ([#3413](https://github.com/dso-toolkit/dso-toolkit/issues/3413))
+* Document Component: alternativeTitle met `href` niet zichtbaar interactief ([#3417](https://github.com/dso-toolkit/dso-toolkit/issues/3417))
+
 ## 💃 Release 81.0.0 - 2025-11-04
 
 ### Added